requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==41.0.7
distlib==0.3.8
dnspython==2.4.2
dulwich==0.21.7
email-validator==2.1.0.post1
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.3
keyring==24.3.0
MarkupSafe==2.1.3
more-itertools==10.2.0
msgpack==1.0.7
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
orjson==3.9.12
packaging==23.2
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyproject_hooks==1.0.0
pytest==7.4.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.0
starlette==0.35.1
tomlkit==0.12.3
trove-classifiers==2024.1.8
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
